Update prev_pc on jmp / call/ ret instructions

+++ src/memory.rs
@@ -53,10 +53,10 @@ impl Memory {
         (self.memory[addr as usize - 1] as u16) >> 8 | (self.memory[addr as usize - 2] as u16)
     }
     pub fn read_word(&mut self, addr: u16) -> u16 {
-        (self.memory[addr as usize + 1] as u16) << 8 | (self.memory[(addr as usize)] as u16)
+        u16::from((self.memory[addr as usize + 1] as u16) << 8 | (self.memory[addr as usize] as u16))
     }
     pub fn write_memory(&mut self, addr: u16) {
-        (self.memory[addr as usize + 2] as u16) << 8 | (self.memory[addr as usize + 1] as u16);
+        u16::from((self.memory[addr as usize + 2] as u16) << 8 | (self.memory[addr as usize + 1] as u16));
     }
 
     pub fn read_high(&mut self, addr: u16) -> u8 {
@@ -75,12 +75,12 @@ impl Memory {
     // Reads the memory address and returns a 16 bit integer, for self.pc / sp
     // instructions
     pub fn read_rp(&mut self, addr: usize) -> u16 {
-        self.memory[addr] as u16
+        u16::from(self.memory[addr])
     }
 
     // Useful to read values out of memory to assign to 8 bit registers
     pub fn read(&mut self, addr: u16) -> u16 {
-        self.memory[addr as usize] as u16
+        u16::from(self.memory[addr as usize])
     }
     pub fn write(&mut self, addr: u16, val: u8) {
         self.memory[addr as usize] = val;
